[fix] Pressable prevent longpress contextmenu on Windows
Long-press with a touch on a Pressable does not suppress the contextmenu
on Windows. On iOS and Android, the contextmenu is triggered while the
pointer is down, after a certain length of time. However, on Windows,
the contextmenu is displayed after pointerup (with the same timestamp as
pointerup), whether triggered by a mouse right-click or a touch
long-press.

Fix #2296
Close #2316