[fix] Make children pressable when parent is disabled
It fixes the inconsistency between native builds and react-native-web,
where when disabling Pressables and other tappable elements, the
children on native builds are still interactable (swiping, tapping, etc)
but un-interactable in react-native-web. On react-native-web, the
pointerEvents were set to 'none' to address 094bd0e, which causes this
inconsistency.

The solution was to change from 'none' to 'box-none', which maintains
the desired behavior of the original issue and addresses the current.

Close #2536
Fix #2391